it is possible to specify "filler" method for property now
Update issue 98
If there is no targetType.writeMethod( Property_Type ) we try to find targetType.writeMethod(getActualArguments[]).
If we succeed we use write method as filler for the property.
For the collection and array (C)  -- C.foreach(writeMethod(_))
For the Map (M) -- M.foreach(writeMethod(key, value))

NB. Examples and tests are still missing....